requirements.txt
Django==1.8
djangorestframework==3.1.1
gunicorn==19.3
argparse==1.3.0
whitenoise==1.0.6
dj-database-url==0.3.0
django-extensions==1.5.2
django-filter==0.9.2
Pillow==2.8.1
psycopg2==2.6
pytz==2014.10
six==1.9.0
mod_wsgi==4.4.11
django-allauth==0.20.0
django-rest-auth==0.4.0